refactoring and more tests
