+++ tests/test_util.py
@@ -1,9 +1,11 @@
 import unittest
+from mock import patch
 
-from lbryum.util import format_satoshis
+from lbryum.util import format_satoshis, user_dir
 
 
-class TestUtil(unittest.TestCase):
+class TestSatoshis(unittest.TestCase):
+
     def test_format_satoshis(self):
         result = format_satoshis(1234)
         expected = "0.00001234"
@@ -18,3 +20,16 @@ class TestUtil(unittest.TestCase):
         result = format_satoshis(-1234, is_diff=True)
         expected = "-0.00001234"
         self.assertEqual(expected, result)
+
+
+class TestUserDir(unittest.TestCase):
+
+    def test_user_dir(self):
+        with patch.dict('os.environ', {'HOME': '/home/joe/'}, clear=True):
+            self.assertEqual('/home/joe/.lbryum', user_dir())
+        with patch.dict('os.environ', {'APPDATA': '/home/joe/'}, clear=True):
+            self.assertEqual('/home/joe/LBRYum', user_dir())
+        with patch.dict('os.environ', {'LOCALAPPDATA': '/home/joe/'}, clear=True):
+            self.assertEqual('/home/joe/LBRYum', user_dir())
+        with patch.dict('os.environ', {}, clear=True):
+            self.assertEqual(None, user_dir())
